requirements.txt
pandas
requests
scikit-learn
joblib
apache-airflow
apache-airflow-providers-http
